Arm4
*Added Arm4 to MattWorkshop, added colliders and components.

As a note, the textured Arm4 doesn't exist in the project right now which I'm assuming happened after the merge(?). The material doesn't exist either so the arm was done on just a non-textured model. The animations from arms 7 and 4 should be already in the existing merge, if that's not the case let me know since it looks good on my end