Arm7/Tail1 Setup
Finished adding animations to the controller for Arm 7
Added animation controller for Tail 1
Added colliders and components to Tail 1
Moved some assets I worked on to folder for organization. If it says that there's a item missing, it was most likely moved instead like material for Tail 1, it was moved.