Tail 1 Anim Updates
Ignore the fact that the branch says it's Arm 1 updates, it meant to say Tail 1

Replaced old Tail 1 animations with new ones Lexi provided. There are some issues with these animations, hoping to talk to Lexi when she's available about it. Specifically the transition between the wind up and heavy attacks.

There are two folders for Tail 1 animations, being the old and new animations. The old ones are staying until the issues get resolved and I can be sure that we don't need them anymore.