Finished Emote System, Teetering System
Emotes are fully functional and work properly with interruption sources, prioritizing all other inputs over emotes. Additionally, the calming system in the idle has been given a lil umph with teetering. Now we can tell the system when we are at a ledge and our character will teeter when it has been left prone for a few seconds