Floating Monsters and Spawning In
Allowed for non winged non grounded leg monsters to just float around like an orb and to have the animations look right. Also in same system allowed players to toggle whether it is a floating character or not in case anyone wanted to move around like dark samus. Last, I added in a spawn animation system that I will be building on more at a later date.