Attack Room Fixes 11-7-24
Brought in the new and improved Monster base that works for both the attack test room as well as the actual player controller. Adjusted the camera transitions and animations to account for the true player size. Basically set Judah up to continue his work with proper animations and a proper scale