Build-a-Scare Test Room
This is an upload of the in-engine stuff I've been doing with Build-a-Scare. I left the main Build-a-Scare scene untouched, I've been working in the Jolly Test Room scene in the Test Rooms folder. All of my changes to the project are only within \Test Rooms\Jolly or \UI\UI Cutouts.