Made Progress with VFX Code
Moved some of the attack specific VFX code into the Neutral attack classes. The plan going forward is that the monster part visual will be for general VFX and visual code and any attack specific code will be called by the monster part visual but will be located in the attack scripts.