Added pages
There are now buttons for going to the next page, going to the previous page, going to the last page and going to the first page. Also added some QOL features like resetting the scroll when changing pages and automatically going to the last page and scrolling to the bottom when adding a new config.