Improved Error Checking and User Feedback
Tried to cover all possible outcomes to ensure the tool doesn't throw an error. Instead of throwing an error the tool will prevent the user from continuing and will inform them of what is missing. I also changed the size and positions of the fields and buttons to make the window look nicer.