Fixed issue where monster part ref was null
After converting the attack scripts to monobehaviors instead of C# scripts there was a bug where the monster part ref was null in the call to the attacks on attack release function. The ref was being passed through the monster part visual script which in some cases was not initialized. To fix it I made it not pass the ref through the visual and just made it use the one from the monster part script which is is initialized in awaken the beast.